+++ qcelemental/models/results.py
@@ -37,8 +37,24 @@ class ResultProperties(BaseModel):
     mp2_opposite_spin_correlation_energy: float = None
     mp2_singles_energy: float = None
     mp2_doubles_energy: float = None
-    mp2_total_correlation_energy: float = None
+    mp2_correlation_energy: float = None
     mp2_total_energy: float = None
+    mp2_dipole_moment: List[float] = None
+
+    # CCSD Keywords
+    ccsd_same_spin_correlation_energy: float = None
+    ccsd_opposite_spin_correlation_energy: float = None
+    ccsd_singles_energy: float = None
+    ccsd_doubles_energy: float = None
+    ccsd_correlation_energy: float = None
+    ccsd_total_energy: float = None
+    ccsd_dipole_moment: List[float] = None
+    ccsd_iterations: int = None
+
+    # CCSD(T) keywords
+    ccsd_prt_pr_correlation_energy: float = None
+    ccsd_prt_pr_total_energy: float = None
+    ccsd_prt_pr_dipole_moment: List[float] = None
 
     class Config:
         allow_mutation = False